+++ Source/ImGui/Private/ImGuiContextProxy.cpp
@@ -6,6 +6,7 @@
 #include "ImGuiDelegatesContainer.h"
 #include "ImGuiImplementation.h"
 #include "ImGuiInteroperability.h"
+#include "Utilities/Arrays.h"
 
 #include <Runtime/Launch/Resources/Version.h>
 
@@ -135,6 +136,7 @@ void FImGuiContextProxy::Tick(float DeltaSeconds)
 		// Update context information (some data, like mouse cursor, may be cleaned in new frame, so we should collect it
 		// beforehand).
 		bHasActiveItem = ImGui::IsAnyItemActive();
+		bIsMouseHoveringAnyWindow = ImGui::IsMouseHoveringAnyWindow();
 		MouseCursor = ImGuiInterops::ToSlateMouseCursor(ImGui::GetMouseCursor());
 		DisplaySize = ImGuiInterops::ToVector2D(ImGui::GetIO().DisplaySize);
 
@@ -150,10 +152,8 @@ void FImGuiContextProxy::BeginFrame(float DeltaTime)
 		ImGuiIO& IO = ImGui::GetIO();
 		IO.DeltaTime = DeltaTime;
 
-		if (InputState)
-		{
-			ImGuiInterops::CopyInput(IO, *InputState);
-		}
+		ImGuiInterops::CopyInput(IO, InputState);
+		InputState.ClearUpdateState();
 
 		ImGui::NewFrame();